Remove duplicate formattable methods and create a module
