+++ app/poros/hourly_forecast.rb
@@ -1,4 +1,6 @@
 class HourlyForecast
+  include Formattable
+  
   attr_reader :time,
               :temperature,
               :wind_speed,
@@ -7,20 +9,12 @@ class HourlyForecast
               :icon
 
   def initialize(hourly_forecast_params)
-    @time = format_time(hourly_forecast_params[:time], '%T')
+    @time = format_strftime_time(hourly_forecast_params[:time], '%T')
     @temperature = hourly_forecast_params[:temperature]
     @wind_speed = wind_speed_to_string(hourly_forecast_params)
     @wind_direction = wind_direction_degrees_to_compass(hourly_forecast_params)
     @conditions = hourly_forecast_params[:conditions]
     @icon = icon_link(hourly_forecast_params[:icon])
-  end
-
-  def format_time(time, format)
-    Time.zone.at(time).strftime(format)
-  end
-
-  def icon_link(icon_code)
-    "http://openweathermap.org/img/wn/#{icon_code}@2x.png"
   end
 
   def wind_speed_to_string(hourly_forecast_params)
